requirements.txt
Pillow>=7.2.0
requests
urllib3
pynacl==1.5.0
orjson>=3.10.0
sv_ttk>=2.6.0
pyinstaller